Refactor services into sidetrack namespace package
