+++ pyproject.toml
@@ -1,3 +1,16 @@
+[project]
+name = "sidetrack"
+version = "0.1.0"
+description = "Multi-service SideTrack application"
+requires-python = ">=3.11"
+
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools.packages.find]
+include = ["sidetrack*"]
+
 [tool.black]
 line-length = 100
 target-version = ["py311"]
