+++ services/api/tests/test_labels.py
@@ -1,19 +1,15 @@
 import os
-import sys
-from pathlib import Path
 
 import pytest
 from fastapi.testclient import TestClient
 from sqlalchemy import select
 
-# Ensure repository root on sys.path and configure SQLite for tests
-sys.path.insert(0, str(Path(__file__).resolve().parents[3]))
-os.environ["DATABASE_URL"] = "sqlite:///./test.db"
+os.environ["DATABASE_URL"] = "sqlite+aiosqlite:///./test.db"
 
-from services.api.app.db import SessionLocal, engine, get_db  # noqa: E402
-from services.api.app.main import app  # noqa: E402  (import after setting env)
-from services.api.app.schemas.labels import LabelResponse
-from services.common.models import Base, Track, UserLabel  # noqa: E402
+from sidetrack.api.db import SessionLocal, engine, get_db  # noqa: E402
+from sidetrack.api.main import app  # noqa: E402  (import after setting env)
+from sidetrack.api.schemas.labels import LabelResponse
+from sidetrack.common.models import Base, Track, UserLabel  # noqa: E402
 
 Base.metadata.create_all(bind=engine)
