+++ sidetrack/api/api/v1/musicbrainz.py
@@ -9,7 +9,7 @@ from fastapi import APIRouter, Depends, HTTPException, Query
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from services.common.models import Artist, Release, Track
+from sidetrack.common.models import Artist, Release, Track
 
 from ...db import get_db
 from ...main import get_http_client
