+++ sidetrack/api/db.py
@@ -7,7 +7,7 @@ from opentelemetry.instrumentation.sqlalchemy import SQLAlchemyInstrumentor
 from sqlalchemy.exc import SQLAlchemyError
 from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker, create_async_engine
 
-from services.common.models import Base
+from sidetrack.common.models import Base
 
 from .config import get_settings
