+++ sidetrack/api/repositories/release_repository.py
@@ -1,6 +1,6 @@
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from services.common.models import Release
+from sidetrack.common.models import Release
 from ..utils import get_or_create
