+++ sidetrack/api/schemas/settings.py
