+++ sidetrack/api/scoring.py
@@ -10,7 +10,7 @@ from pathlib import Path
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from services.common.models import Embedding, Feature
+from sidetrack.common.models import Embedding, Feature
 
 from .constants import AXES, SUPPORTED_METHODS
